+++ build.washingtonpost.com/snippets/avatar/avatar.jsx
@@ -10,7 +10,7 @@ export default function Example() {
   return (
     <Avatar size="400">
       <img
-        src="https://i.pravatar.cc/300"
+        src="https://i.pravatar.cc/300/300"
         alt="An avatar is an atomic component that represents an individual’s identity through a circular photo."
       />
     </Avatar>