+++ ui/input-text/src/InputText.tsx
@@ -1,4 +1,5 @@
-import React from "react";
+import * as React from "react";
+import { useEffect, useState } from "react";
 import { nanoid } from "nanoid";
 import { theme, styled } from "@washingtonpost/wpds-theme";
 import { Button } from "@washingtonpost/wpds-button";
@@ -6,6 +7,7 @@ import { Icon } from "@washingtonpost/wpds-icon";
 import {
   useFloating,
   unstyledInputStyles,
+  globalInputAutoFillTriggerAnimations,
 } from "@washingtonpost/wpds-input-shared";
 import { InputLabel } from "@washingtonpost/wpds-input-label";
 import { ErrorMessage } from "@washingtonpost/wpds-error-message";
@@ -96,6 +98,8 @@ export interface InputTextProps
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
   /** Callback executed when the input fires a focus event */
   onFocus?: React.FocusEventHandler<HTMLInputElement>;
+  /** placeholder text */
+  placeholder?: string;
   /** The input elements required attribute */
   required?: boolean;
   /** indicates there is a success */
@@ -126,6 +130,7 @@ export const InputText = React.forwardRef<HTMLInputElement, InputTextProps>(
       onChange,
       onFocus,
       onButtonIconClick,
+      placeholder,
       required,
       success,
       type = "text",
@@ -135,16 +140,54 @@ export const InputText = React.forwardRef<HTMLInputElement, InputTextProps>(
     },
     ref
   ) => {
-    const [helperId, setHelperId] = React.useState<string | undefined>();
-    const [errorId, setErrorId] = React.useState<string | undefined>();
+    const [helperId, setHelperId] = useState<string | undefined>();
+    const [errorId, setErrorId] = useState<string | undefined>();
+    const [isAutofilled, setIsAutofilled] = useState<boolean>(false);
+    const internalRef = React.useRef<HTMLInputElement>(null);
 
-    React.useEffect(() => {
+    useEffect(() => {
       setHelperId(`wpds-input-helper-${nanoid(6)}`);
       setErrorId(`wpds-input-error-${nanoid(6)}`);
     }, []);
 
+    //takes into account ref that might be passed into the component
+    useEffect(() => {
+      if (!ref) return;
+
+      if (typeof ref === "function") {
+        ref(internalRef.current);
+      } else {
+        ref.current = internalRef.current;
+      }
+    }, [ref, internalRef]);
+
+    useEffect(() => {
+      globalInputAutoFillTriggerAnimations();
+      const element = internalRef.current;
+
+      const onAnimationStart = (e) => {
+        switch (e.animationName) {
+          case "jsTriggerAutoFillStart":
+            return setIsAutofilled(true);
+          case "jsTriggerAutoFillCancel":
+            return setIsAutofilled(false);
+        }
+      };
+
+      element?.addEventListener("animationstart", onAnimationStart, false);
+
+      return () => {
+        element?.removeEventListener("animationstart", onAnimationStart, false);
+      };
+    }, []);
+
     const [isFloating, handleOnFocus, handleOnBlur, handleOnChange] =
-      useFloating(value || defaultValue, onFocus, onBlur, onChange);
+      useFloating(
+        value || defaultValue || placeholder || isAutofilled,
+        onFocus,
+        onBlur,
+        onChange
+      );
 
     function handleButtonIconClick(event) {
       onButtonIconClick && onButtonIconClick(event);
@@ -238,7 +281,8 @@ export const InputText = React.forwardRef<HTMLInputElement, InputTextProps>(
               onBlur={handleOnBlur}
               onChange={handleOnChange}
               onFocus={handleOnFocus}
-              ref={ref}
+              placeholder={placeholder}
+              ref={internalRef}
               required={required}
               type={type}
               value={value}