+++ ui/input-textarea/src/InputTextarea.tsx
@@ -1,11 +1,13 @@
 import * as React from "react";
+import { useEffect, useState } from "react";
 import { nanoid } from "nanoid";
-import * as Theme from "@washingtonpost/wpds-theme";
+import { theme, css, styled } from "@washingtonpost/wpds-theme";
 import type * as WPDS from "@washingtonpost/wpds-theme";
 import {
   sharedInputStyles,
   sharedInputVariants,
   useFloating,
+  globalInputAutoFillTriggerAnimations,
 } from "@washingtonpost/wpds-input-shared";
 import { InputLabel } from "@washingtonpost/wpds-input-label";
 import { ErrorMessage } from "@washingtonpost/wpds-error-message";
@@ -13,7 +15,7 @@ import { HelperText } from "@washingtonpost/wpds-helper-text";
 
 const NAME = "InputTextarea";
 
-const InputTextareaCSS = Theme.css({
+const InputTextareaCSS = css({
   ...sharedInputStyles,
   display: "block",
   minHeight: "$500",
@@ -33,23 +35,23 @@ const InputTextareaCSS = Theme.css({
   },
 });
 
-const TextAreaLabel = Theme.styled(InputLabel, {
+const TextAreaLabel = styled(InputLabel, {
   insetBlockStart: "$050",
   insetInlineStart: "$050",
   pointerEvents: "none",
   position: "absolute",
-  transition: Theme.theme.transitions.allFast,
+  transition: theme.transitions.allFast,
   variants: {
     isFloating: {
       true: {
-        fontSize: Theme.theme.fontSizes["075"],
-        lineHeight: Theme.theme.lineHeights["100"],
+        fontSize: theme.fontSizes["075"],
+        lineHeight: theme.lineHeights["100"],
       },
     },
   },
 });
 
-const ControlCSS = Theme.css({
+const ControlCSS = css({
   display: "flex",
   flexDirection: "column",
   position: "relative",
@@ -84,6 +86,8 @@ export interface InputTextareaProps
   onChange?: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
   /** Callback executed when the input fires a focus event */
   onFocus?: React.FocusEventHandler<HTMLTextAreaElement>;
+  /** placeholder text */
+  placeholder?: string;
   /** The input elements required attribute */
   required?: boolean;
   /** The input element value for controlled components */
@@ -112,20 +116,55 @@ export const InputTextarea = React.forwardRef<
       onBlur,
       onChange,
       required,
+      placeholder,
       ...props
     },
     ref
   ) => {
-    const [helperId, setHelperId] = React.useState<string | undefined>();
-    const [errorId, setErrorId] = React.useState<string | undefined>();
+    const [helperId, setHelperId] = useState<string | undefined>();
+    const [errorId, setErrorId] = useState<string | undefined>();
+    const [isAutofilled, setIsAutofilled] = useState<boolean>(false);
 
-    React.useEffect(() => {
+    const internalRef = React.useRef<HTMLTextAreaElement>(null);
+
+    useEffect(() => {
       setHelperId(`wpds-input-helper-${nanoid(6)}`);
       setErrorId(`wpds-input-error-${nanoid(6)}`);
     }, []);
 
+    //takes into account ref that might be passed into the component
+    useEffect(() => {
+      if (!ref) return;
+
+      if (typeof ref === "function") {
+        ref(internalRef.current);
+      } else {
+        ref.current = internalRef.current;
+      }
+    }, [ref, internalRef]);
+
+    useEffect(() => {
+      globalInputAutoFillTriggerAnimations();
+      const element = internalRef.current;
+
+      const onAnimationStart = (e) => {
+        switch (e.animationName) {
+          case "jsTriggerAutoFillStart":
+            return setIsAutofilled(true);
+          case "jsTriggerAutoFillCancel":
+            return setIsAutofilled(false);
+        }
+      };
+
+      element?.addEventListener("animationstart", onAnimationStart, false);
+
+      return () => {
+        element?.removeEventListener("animationstart", onAnimationStart, false);
+      };
+    }, []);
+
     const [isFloating, handleFocus, handleBlur, handleChange] = useFloating(
-      value || defaultValue,
+      value || defaultValue || placeholder || isAutofilled,
       onFocus,
       onBlur,
       onChange
@@ -137,14 +176,16 @@ export const InputTextarea = React.forwardRef<
           {...props}
           id={id}
           name={name}
-          ref={ref}
+          ref={internalRef}
           className={InputTextareaCSS({
             css: css,
             canResize: canResize,
             isInvalid: error,
             isDisabled: disabled,
           })}
+          required={required}
           disabled={disabled}
+          placeholder={placeholder}
           onFocus={handleFocus}
           onBlur={handleBlur}
           onChange={handleChange}