+++ ui/select/CHANGELOG.md
@@ -3,13 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.20.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.20.0...v0.20.1) (2022-10-24)
+# [0.24.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.23.2...v0.24.0) (2023-02-03)
 
 **Note:** Version bump only for package @washingtonpost/wpds-select
 
 
 
 
+
+## [0.23.2](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.24.0...v0.23.2) (2023-01-17)
+
+**Note:** Version bump only for package @washingtonpost/wpds-select
+
+
+
+
+
+## [0.23.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.23.0...v0.23.1) (2023-01-11)
+
+**Note:** Version bump only for package @washingtonpost/wpds-select
+
+
+
+
+
+# [0.23.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.22.0...v0.23.0) (2023-01-11)
+
+**Note:** Version bump only for package @washingtonpost/wpds-select
+
+# [0.22.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.21.0...v0.22.0) (2023-01-05)
+
+### Features
+
+- allow for use of Select in example form on mobile ([#288](https://github.com/washingtonpost/wpds-ui-kit/issues/288)) ([72291c6](https://github.com/washingtonpost/wpds-ui-kit/commit/72291c65dc51d5627e0d3f18fd816637eb29c523))
+
+# [0.21.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.20.1...v0.21.0) (2022-12-02)
+
+**Note:** Version bump only for package @washingtonpost/wpds-select
+
+## [0.20.1](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.20.0...v0.20.1) (2022-10-24)
+
+**Note:** Version bump only for package @washingtonpost/wpds-select
 
 # [0.20.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.19.1...v0.20.0) (2022-10-21)